fix: pr code style feedback
