+++ src/server/resetServerUnistyles.ts
@@ -1,0 +1,9 @@
+import { error, isServer } from '../web/utils'
+import { UnistylesWeb } from '../web'
+
+export const resetServerUnistyles = () => {
+    if (!isServer()) {
+        throw error('Server styles should only be reset on the server')
+    }
+    UnistylesWeb.registry.reset()
+}
