+++ src/server/types.ts
@@ -1,0 +1,4 @@
+export type ServerUnistylesSettings = {
+  includeRNWStyles?: boolean
+}
+export const DefaultServerUnistylesSettings = {}
