+++ src/server/useServerUnistyles.tsx
@@ -1,45 +1,16 @@
 import React, { useRef } from 'react'
-import { StyleSheet } from 'react-native'
-import { UnistylesWeb } from '../web'
 import { isServer } from '../web/utils'
+import { DefaultServerUnistylesSettings, type ServerUnistylesSettings } from './types'
+import { getServerUnistyles } from './getServerUnistyles'
+import { resetServerUnistyles } from './resetServerUnistyles'
+import { hydrateServerUnistyles } from './hydrateServerUnistyles'
 
-declare global {
-    interface Window {
-        // @ts-ignore
-        __UNISTYLES_STATE__: ReturnType<typeof UnistylesWeb.registry.css.getState>
-    }
-}
-
-export const getServerUnistyles = (includeRNWStyles = true) => {
-    if (!isServer()) throw new Error('Server styles should only be read on the server')
-    // @ts-ignore
-    const rnwStyle: string | null = includeRNWStyles ? (StyleSheet?.getSheet().textContent ?? '') : null
-    const css = UnistylesWeb.registry.css.getStyles()
-    const state = UnistylesWeb.registry.css.getState()
-    return <>
-        {rnwStyle && <style id='rnw-style'>{rnwStyle}</style>}
-        <style id='unistyles-web'>{css}</style>
-        <script id='unistyles-script'>{`window.__UNISTYLES_STATE__ = ${JSON.stringify(state)}`}</script>
-    </>
-}
-
-export const resetServerUnistyles = () => {
-    if (!isServer()) throw new Error('Server styles should only be reset on the server')
-    UnistylesWeb.registry.reset()
-}
-
-export const hydrateServerUnistyles = () => {
-    if (isServer()) throw new Error('Server styles should only be hydrated on the client')
-    UnistylesWeb.registry.css.hydrate(window.__UNISTYLES_STATE__)
-    document.querySelector('#unistyles-script')?.remove()
-}
-
-export const useServerUnistyles = (includeRNWStyles = true): React.ReactNode | null => {
+export const useServerUnistyles = (settings: ServerUnistylesSettings = DefaultServerUnistylesSettings): React.ReactNode | null => {
     const isServerInserted = useRef(false)
 
     if (isServer() && !isServerInserted.current) {
         isServerInserted.current = true
-        const components = getServerUnistyles(includeRNWStyles)
+        const components = getServerUnistyles(settings)
         resetServerUnistyles()
         return components
     }
